Added diagram pdf, refactored interface package.
