+++ src/main/java/io/exam/animal/model/FlyingInsect.java
@@ -1,6 +1,7 @@
-package io.exam.animal.model.impl;
+package io.exam.animal.model;
 
 import io.exam.animal.model.FlyingAnimal;
+import io.exam.animal.model.Insect;
 
 public interface FlyingInsect extends Insect, FlyingAnimal {
 }
