+++ src/main/java/io/exam/animal/model/Insect.java
@@ -1,0 +1,5 @@
+package io.exam.animal.model;
+
+public interface Insect extends WalkingAnimal {
+
+}
