--- src/main/java/io/exam/animal/model/impl/CrawlingInsect.java
@@ -1,6 +1,0 @@
-package io.exam.animal.model.impl;
-
-import io.exam.animal.model.WalkingAnimal;
-
-public interface CrawlingInsect extends Insect, WalkingAnimal {
-}
