--- src/main/java/io/exam/animal/model/impl/Insect.java
@@ -1,5 +1,0 @@
-package io.exam.animal.model.impl;
-
-public interface Insect {
-
-}
