+++ src/test/java/test/io/exam/animal/model/TestInsect.java
@@ -1,11 +1,8 @@
 package test.io.exam.animal.model;
 
-import io.exam.animal.model.FlyingAnimal;
-import io.exam.animal.model.WalkingAnimal;
+import io.exam.animal.model.*;
 import io.exam.animal.model.impl.Butterfly;
 import io.exam.animal.model.impl.Caterpillar;
-import io.exam.animal.model.impl.CrawlingInsect;
-import io.exam.animal.model.impl.FlyingInsect;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -26,7 +23,7 @@ public class TestInsect {
     void testButterflyCanAlsoCrawl() {
 
         Butterfly butterfly = new Butterfly();
-        assertTrue(butterfly instanceof CrawlingInsect);
+        assertTrue(butterfly instanceof Insect);
         assertTrue(butterfly instanceof WalkingAnimal);
 
         assertEquals("I am crawling", butterfly.walk());
@@ -36,7 +33,7 @@ public class TestInsect {
     void testCaterpillarCanNotFly() {
 
         Caterpillar caterpillar = new Caterpillar();
-        assertTrue(caterpillar instanceof CrawlingInsect);
+        assertTrue(caterpillar instanceof Insect);
         assertTrue(caterpillar instanceof WalkingAnimal);
         assertFalse(caterpillar instanceof FlyingInsect);
         assertFalse(caterpillar instanceof FlyingAnimal);
